Change ruby_version scheme
This changeset changes the version scheme for the Ruby installation from
package version numbers to Ruby version number.

Whenever a new package is released for Ruby in the PPA used by this
role, it doesn't get added to the repository, instead it replaces the
existing package. This becomes problematic if you are attempting to pin
package versions.